backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
langchain
openai
tiktoken
python-multipart
pytest
pytest-asyncio
python-dotenv
redis
pydantic-settings
pypdf
langchain-community
numpy
python-json-logger==2.0.7
google-generativeai
